\name{RCDat}
\alias{RCDat}
\docType{data}

\title{A simulated dataset for random coefficient models.} 

\description{
Number of customers I=100 and each customer responds to J=10 items. The first column is for responses. It should be used in \code{RCLM(RCDat)}.
\describe{
  \item{ RCDat }{ A simulated dataset.}
}
}
\usage{data(RCDat)}
\format{matrix}




\references{
Liang, F., Jia, B., Xue, J., Li, Q., and Luo, Y. (2017). An Imputation-Consistency Algorithm for High-Dimensional Missing Data Problems and Beyond. Submitted to Journal of the Royal Statistical Society Series B.
}


\keyword{datasets}
